refactor: add `zimo` namespace to `mw`
