+++ src/app_main.cpp
@@ -31,12 +31,12 @@
 #include "intf/usb/init.hpp"
 #include "mem/nvs/init.hpp"
 #include "mw/dcc/init.hpp"
-#include "mw/decup/init.hpp"
-#include "mw/mdu/init.hpp"
 #include "mw/ota/init.hpp"
-#include "mw/ulf/init.hpp"
 #include "mw/z21/init.hpp"
-#include "mw/zusi/init.hpp"
+#include "mw/zimo/decup/init.hpp"
+#include "mw/zimo/mdu/init.hpp"
+#include "mw/zimo/ulf/init.hpp"
+#include "mw/zimo/zusi/init.hpp"
 #include "utility.hpp"
 
 /// ESP-IDF application entry point
@@ -64,24 +64,24 @@ extern "C" void app_main() {
   ESP_ERROR_CHECK(invoke_on_core(PRO_CPU_NUM, intf::udp::init));
   ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::dcc::init));
   static_assert(APP_CPU_NUM == mw::dcc::task.core_id);
-  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::decup::init));
-  static_assert(APP_CPU_NUM == mw::decup::task.core_id);
-  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::mdu::init));
-  static_assert(APP_CPU_NUM == mw::mdu::task.core_id);
   ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::ota::init));
   static_assert(APP_CPU_NUM == mw::ota::task.core_id);
   ESP_ERROR_CHECK(invoke_on_core(PRO_CPU_NUM, mw::z21::init));
   static_assert(APP_CPU_NUM == mw::z21::task.core_id);
-  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::zusi::init));
-  static_assert(APP_CPU_NUM == mw::zusi::task.core_id);
+  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::zimo::zusi::init));
+  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::zimo::decup::init));
+  static_assert(APP_CPU_NUM == mw::zimo::decup::task.core_id);
+  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::zimo::mdu::init));
+  static_assert(APP_CPU_NUM == mw::zimo::mdu::task.core_id);
+  static_assert(APP_CPU_NUM == mw::zimo::zusi::task.core_id);
   ESP_ERROR_CHECK(invoke_on_core(PRO_CPU_NUM, intf::mdns::init));
 
   // Don't disable serial JTAG
 #if !defined(CONFIG_USJ_ENABLE_USB_SERIAL_JTAG)
-  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::ulf::init));
-  static_assert(APP_CPU_NUM == mw::ulf::dcc_ein::task.core_id &&
-                APP_CPU_NUM == mw::ulf::decup_ein::task.core_id &&
-                APP_CPU_NUM == mw::ulf::susiv2::task.core_id);
+  ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, mw::zimo::ulf::init));
+  static_assert(APP_CPU_NUM == mw::zimo::ulf::dcc_ein::task.core_id &&
+                APP_CPU_NUM == mw::zimo::ulf::decup_ein::task.core_id &&
+                APP_CPU_NUM == mw::zimo::ulf::susiv2::task.core_id);
   ESP_ERROR_CHECK(invoke_on_core(APP_CPU_NUM, intf::usb::init));
   static_assert(APP_CPU_NUM == intf::usb::rx_task.core_id &&
                 APP_CPU_NUM == intf::usb::tx_task.core_id);
@@ -100,14 +100,14 @@ static_assert(std::invoke([] {
                         intf::usb::rx_task.name,
                         intf::usb::tx_task.name,
                         mw::dcc::task.name,
-                        mw::decup::task.name,
-                        mw::mdu::task.name,
                         mw::ota::task.name,
-                        mw::ulf::dcc_ein::task.name,
-                        mw::ulf::decup_ein::task.name,
-                        mw::ulf::susiv2::task.name,
                         mw::z21::task.name,
-                        mw::zusi::task.name};
+                        mw::zimo::decup::task.name,
+                        mw::zimo::mdu::task.name,
+                        mw::zimo::ulf::dcc_ein::task.name,
+                        mw::zimo::ulf::decup_ein::task.name,
+                        mw::zimo::ulf::susiv2::task.name,
+                        mw::zimo::zusi::task.name};
   std::ranges::sort(task_names, [](char const* a, char const* b) {
     return ztl::strcmp(a, b) < 0;
   });
