+++ src/drv/analog/adc_task_function.cpp
@@ -56,7 +56,7 @@ auto get_short_circuit_count() {
 ///
 /// If the measured currents indicate a short circuit, the \ref led::bug
 /// "bug LED" is switched on, \ref state is set to \ref State::ShortCircuit
-/// "short circuit" and a \ref page_z21 track short circuit message is
+/// "short circuit" and a \ref page_mw_z21 track short circuit message is
 /// broadcast.
 [[noreturn]] void adc_task_function(void*) {
   std::array<uint8_t, conversion_frame_size> conversion_frame{};
