+++ src/intf/doxygen.hpp
@@ -25,23 +25,26 @@ namespace intf {
 
 /// \page page_intf Interfaces
 /// \details
+/// The interface modules contain the connections to the outside world.
+/// Currently, these are lwIP (over WiFi), UDP, and USB.
+///
 // clang-format off
 /// \page page_intf Interface
 /// \details
-/// | Chapter                | Namespace | Content                                                                                                                                                                                                                           |
-/// | ---------------------- | --------- | ---------------------------------------------------------------------------------- |
-/// | \subpage page_http     | \ref http | Access point (AP) and station (STA) HTTP servers                                   |
-/// | \subpage page_mdns     | \ref mdns | mDNS services                                                                      |
-/// | \subpage page_udp      | \ref udp  | Create and bind UDP socket                                                         |
-/// | \subpage page_usb      | \ref usb  | Create [CDC](https://en.wikipedia.org/wiki/USB_communications_device_class) device |
+/// | Chapter                 | Namespace | Content                                                                                                                                                                                                                           |
+/// | ----------------------- | --------- | ---------------------------------------------------------------------------------- |
+/// | \subpage page_intf_http | \ref http | Access point (AP) and station (STA) HTTP servers                                   |
+/// | \subpage page_intf_mdns | \ref mdns | Register mDNS services                                                             |
+/// | \subpage page_intf_udp  | \ref udp  | Create and bind UDP socket                                                         |
+/// | \subpage page_intf_usb  | \ref usb  | Create [CDC](https://en.wikipedia.org/wiki/USB_communications_device_class) device |
 // clang-format on
 /// \page page_intf Interface
 /// \details
 ///
 /// <div class="section_buttons">
-/// | Previous                | Next           |
-/// | :---------------------- | -------------: |
-/// | \ref page_api_reference | \ref page_http |
+/// | Previous                | Next                |
+/// | :---------------------- | ------------------: |
+/// | \ref page_api_reference | \ref page_intf_http |
 /// </div>
 
 } // namespace intf
