+++ src/intf/usb/doxygen.hpp
@@ -23,25 +23,25 @@
 
 namespace intf::usb {
 
-/// \page page_usb USB
-/// \tableofcontents
+/// \page page_intf_usb USB
+/// \details \tableofcontents
 /// This module initializes the receive and transmit routines of the USB
 /// interface. Currently only ZIMO protocols are supported via USB, but an
 /// extension is conceivable.
 ///
-/// \section section_usb_init Initialization
+/// \section section_intf_usb_init Initialization
 /// \copydetails init
 ///
-/// \section section_usb_receive Receive task
+/// \section section_intf_usb_receive Receive task
 /// \copydetails rx_task_function
 ///
-/// \section section_usb_transmit Transmit task
+/// \section section_intf_usb_transmit Transmit task
 /// \copydetails tx_task_function
 ///
 /// <div class="section_buttons">
-/// | Previous      | Next         |
-/// | :------------ | -----------: |
-/// | \ref page_udp | \ref page_mw |
+/// | Previous           | Next         |
+/// | :----------------- | -----------: |
+/// | \ref page_intf_udp | \ref page_mw |
 /// </div>
 
 } // namespace intf::usb
