+++ src/mw/zimo/decup/service.cpp
@@ -14,34 +14,43 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 #include "service.hpp"
-#include <ulf/susiv2.hpp>
 #include <ztl/utility.hpp>
+#include <zusi/zusi.hpp>
 #include "drv/led/bug.hpp"
 #include "log.h"
 #include "utility.hpp"
 
-namespace mw::zusi {
+namespace mw::zimo::decup {
 
-using ::ulf::susiv2::ack, ::ulf::susiv2::nak;
+using ::ulf::decup_ein::ack, ::ulf::decup_ein::nak;
 
 /// \todo document
 Service::Service() {
   task.function = ztl::make_trampoline(this, &Service::taskFunction);
 }
 
-/// \todo document
 /// \bug should this broadcast Z21 programming mode?
-esp_err_t Service::socket(intf::http::Message& msg) {
+esp_err_t Service::zppSocket(intf::http::Message& msg) {
+  return socket(msg, State::DECUPZpp);
+}
+
+/// \bug should this broadcast Z21 programming mode?
+esp_err_t Service::zsuSocket(intf::http::Message& msg) {
+  return socket(msg, State::DECUPZsu);
+}
+
+/// \bug should this broadcast Z21 programming mode?
+esp_err_t Service::socket(intf::http::Message& msg, State decup_state) {
   //
   if (auto expected{State::Suspended};
       msg.type != HTTPD_WS_TYPE_CLOSE &&
-      state.compare_exchange_strong(expected, State::ZUSI)) {
+      state.compare_exchange_strong(expected, decup_state)) {
     _queue.push(std::move(msg));
-    LOGI_TASKS_CREATE(task, drv::out::susi::zimo::zusi::task);
+    LOGI_TASKS_CREATE(task, drv::out::track::zimo::decup::task);
     return ESP_OK;
   }
   //
-  else if (state.load() == State::ZUSI) {
+  else if (state.load() == decup_state) {
     _queue.push(std::move(msg));
     return ESP_OK;
   }
@@ -53,7 +62,8 @@ esp_err_t Service::socket(intf::http::Message& msg) {
 /// \todo document
 [[noreturn]] void Service::taskFunction(void*) {
   switch (state.load()) {
-    case State::ZUSI: loop(); break;
+    case State::DECUPZpp: [[fallthrough]];
+    case State::DECUPZsu: loop(); break;
     default: assert(false); break;
   }
   LOGI_TASK_DESTROY();
@@ -62,7 +72,6 @@ esp_err_t Service::socket(intf::http::Message& msg) {
 /// \todo document
 void Service::loop() {
   drv::led::Bug const led_bug{};
-  auto const timeout{http_receive_timeout2ms()};
 
   for (;;) {
     assert(_queue.size());
@@ -70,25 +79,28 @@ void Service::loop() {
     _queue.pop();
 
     switch (msg.type) {
-      case HTTPD_WS_TYPE_BINARY: _resp = transmit(msg.payload); break;
+      case HTTPD_WS_TYPE_BINARY:
+        std::ranges::for_each(msg.payload,
+                              [&](uint8_t byte) { _ack = receive(byte); });
+        break;
       case HTTPD_WS_TYPE_CLOSE: LOGI("WebSocket closed"); return close();
       default:
         LOGE("WebSocket packet type neither binary nor close");
-        _resp.front() = nak;
-        _resp.resize(sizeof(nak));
+        _ack = nak;
         break;
     }
 
     if (auto const err{httpd_queue_work(new intf::http::Message{
           .sock_fd = msg.sock_fd,
           .type = HTTPD_WS_TYPE_BINARY,
-          .payload = {cbegin(_resp), cend(_resp)},
+          .payload =
+            _ack ? std::vector<uint8_t>{*_ack} : std::vector<uint8_t>{},
         })}) {
       LOGE("httpd_queue_work failed %s", esp_err_to_name(err));
       return close();
     }
 
-    TickType_t const then{xTaskGetTickCount() + pdMS_TO_TICKS(timeout)};
+    TickType_t const then{xTaskGetTickCount() + pdMS_TO_TICKS(task.timeout)};
     while (empty(_queue))
       if (xTaskGetTickCount() >= then) {
         LOGI("WebSocket timeout");
@@ -100,46 +112,23 @@ void Service::loop() {
 }
 
 /// \todo document
-::ulf::susiv2::Response
-Service::transmit(std::vector<uint8_t> const& payload) const {
-  //
-  xMessageBufferSend(drv::out::tx_message_buffer.front_handle,
-                     data(payload),
-                     std::min(size(payload), ZUSI_MAX_PACKET_SIZE),
-                     portMAX_DELAY);
-
-  //
-  decltype(_resp) retval;
-  auto const bytes_received{
-    xMessageBufferReceive(drv::out::rx_message_buffer.handle,
-                          data(retval),
-                          retval.max_size(),
-                          portMAX_DELAY)};
-  retval.resize(bytes_received);
-
-  return retval;
+uint8_t Service::transmit(std::span<uint8_t const> bytes) {
+  uint8_t acks{};
+  if (!xMessageBufferSend(drv::out::tx_message_buffer.front_handle,
+                          data(bytes),
+                          std::min(size(bytes), DECUP_MAX_PACKET_SIZE),
+                          portMAX_DELAY))
+    return acks;
+  else
+    xMessageBufferReceive(
+      drv::out::rx_message_buffer.handle, &acks, sizeof(acks), portMAX_DELAY);
+  return acks;
 }
 
 /// \todo document
 void Service::close() {
   _queue = {};
-
-  // send exit command... just in case?
-  // std::array<uint8_t, 10uz> exit_cmd{};
-  // exit_cmd[5uz] = std::to_underlying(Command::Exit);
-  // exit_cmd[6uz] = 0x55u;
-  // exit_cmd[7uz] = 0xAAu;
-  // exit_cmd[8uz] = 0xFFu;  // No reboot, no CV8 reset
-  // exit_cmd[9uz] = crc8({cbegin(exit_cmd), 9uz});
-  // xMessageBufferSend(out::tx_message_buffer.front_handle,
-  //                    data(exit_cmd),
-  //                    size(exit_cmd),
-  //                    portMAX_DELAY);
-
-  /// \todo wait for ZUSI task to suspend itself?
-
-  // auto expected{State::ZUSI};
-  // state.compare_exchange_strong(expected, State::Suspend);
+  state.store(State::Suspend);
 }
 
-} // namespace mw::zusi
+} // namespace mw::zimo::decup
