+++ src/mw/zimo/mdu/service.cpp
@@ -20,7 +20,7 @@
 #include "log.h"
 #include "utility.hpp"
 
-namespace mw::mdu {
+namespace mw::zimo::mdu {
 
 using ::ulf::mdu_ein::ack, ::ulf::mdu_ein::nak;
 
@@ -131,4 +131,4 @@ Service::transmit(std::vector<uint8_t> const& payload) const {
 /// \todo document
 void Service::close() { _queue = {}; }
 
-} // namespace mw::mdu
+} // namespace mw::zimo::mdu
